# docs/architecture-pack/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Farmmap API
  version: 1.0.0
  description: >
    Farmmap public and authenticated API covering all Must-Have features (F1–F16).
    All endpoints require a Supabase JWT Bearer token where marked. directory_id
    is resolved server-side from the Host header; it is never a client parameter
    on public endpoints.
  contact:
    name: Farmmap Engineering
    url: https://farmmap.co.uk
servers:
  - url: https://farmmap.co.uk/api
    description: Production
  - url: https://staging.farmmap.co.uk/api
    description: Staging
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from auth.getSession()
  parameters:
    ListingSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        example: hargreaves-farm-shop
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: Listing not found
        details:
          type: object
          additionalProperties: true
...
    SubscriptionTier:
      type: string
      enum: [free, bronze, silver, gold]
...
    Product:
      type: object
      required: [id, name, category, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        price_pence:
          type: integer
          nullable: true
        price_currency:
          type: string
          enum: [GBP, EUR]
          default: GBP
        price_display:
          type: string
          maxLength: 100
        category:
          type: string
        allergens_contain:
          type: array
          items:
            type: string
            enum:
              - celery
              - cereals_containing_gluten
              - crustaceans
              - eggs
              - fish
              - lupin
              - milk
              - molluscs
              - mustard
              - peanuts
              - sesame
              - soybeans
              - sulphur_dioxide_sulphites
              - tree_nuts
          nullable: true
        allergens_may_contain:
          type: string
          nullable: true
        allergen_declaration_complete:
          type: boolean
          readOnly: true
        is_purchasable:
          type: boolean
        is_perishable:
          type: boolean
          nullable: true
        is_seasonal:
          type: boolean
        seasonal_months:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 12
          nullable: true
        stock_count:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, draft, archived]
        photo_url:
          type: string
          format: uri
          nullable: true
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /listings/{slug}/products:
    get:
      operationId: listProducts
      summary: List products for a listing
      description: >
        Public endpoint. Returns active products. For Bronze listings, returns
        display-only products. For Silver/Gold, returns purchasable products
        with allergen information. Allergen data is always included.
      tags: [Products]
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
        - name: category
          in: query
          schema:
            type: string
        - name: is_purchasable
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  tier:
                    $ref: '#/components/schemas/SubscriptionTier'

    post:
      operationId: createProduct
      summary: Create a product
      description: >
        Creates a new product (draft status by default). Requires owner or manager
        role and Bronze+ subscription. is_purchasable may only be set to true when
        allergen_declaration_complete is true and the listing is Silver/Gold.
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: is_purchasable requested without allergen data complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /listings/{slug}/products/{productId}:
    put:
      operationId: updateProduct
      summary: Update a product
      description: Requires owner or manager role and Bronze+ subscription.
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error (e.g., setting purchasable without allergen data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteProduct
      summary: Archive a product
      description: Sets product status to archived. Requires owner or manager role.
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingSlug'
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Product archived
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # ORDERS (F8 Silver/Gold marketplace)
  # ============================================================
...